requirements.txt
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==24.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
